refactor: remove unused services and common files; add initial entity models for Inventory, Customer, Employee, and AuditTrail
