+++ Source/RecordManagementApp/RecordManagementAppAPIService/pom.xml
@@ -68,11 +68,6 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>com.danacore</groupId>
-			<artifactId>recordmanagement-common</artifactId>
-			<version>0.0.1-SNAPSHOT</version>
-		</dependency>
-		<dependency>
 			<groupId>com.h2database</groupId>
 			<artifactId>h2</artifactId>
 			<scope>test</scope>
