+++ Source/RecordManagementApp/RecordManagementAppAPIService/src/main/java/com/danacore/recordmanagement/repository/AuditTrailRepository.java
@@ -1,6 +1,6 @@
 package com.danacore.recordmanagement.repository;
 
-import com.danacore.recordmanagement.common.model.AuditTrail;
+import com.danacore.recordmanagement.model.AuditTrail;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
