+++ Source/RecordManagementApp/RecordManagementAppAPIService/src/main/java/com/danacore/recordmanagement/repository/EmployeeRepository.java
@@ -1,6 +1,6 @@
 package com.danacore.recordmanagement.repository;
 
-import com.danacore.recordmanagement.common.model.Employee;
+import com.danacore.recordmanagement.model.Employee;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
