+++ Source/RecordManagementApp/RecordManagementAppAPIService/src/main/java/com/danacore/recordmanagement/repository/InventoryRepository.java
@@ -1,6 +1,6 @@
 package com.danacore.recordmanagement.repository;
 
-import com.danacore.recordmanagement.common.model.Inventory;
+import com.danacore.recordmanagement.model.Inventory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
